backend/requirements.txt
fastapi
uvicorn[standard]
requests
pandas
numpy
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-firestore
google-auth